requirements.txt
amqp==5.0.2; python_version >= "3.6"
beautifulsoup4==4.9.3; python_version >= "3.7"
bencodepy==0.9.5
billiard==3.6.3.0; python_version >= "3.6"
blurhash==1.1.4
celery==5.0.2; python_version >= "3.6"
certifi==2020.11.8; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
chardet==3.0.4; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
click-didyoumean==0.0.3; python_version >= "3.6"
click-repl==0.1.6; python_version >= "3.6"
click==7.1.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
configparser==5.0.1; python_version >= "3.6"
decorator==4.4.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.2.0"
ffmpeg-python==0.2.0
future==0.18.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
idna==2.10; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
importlib-metadata==3.1.0; python_version < "3.8" and python_version >= "3.6"
iso8601==0.1.13; python_version >= "3.7"
kombu==5.0.2; python_version >= "3.6"
lxml==4.6.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
m3u8==0.7.1; python_version >= "3.5"
mastodon.py==1.5.1
minio==6.0.0
numpy==1.19.4; python_version >= "3.6"
pillow==8.0.1; python_version >= "3.6"
prompt-toolkit==3.0.8; python_full_version >= "3.6.1" and python_version >= "3.6"
pycryptodomex==3.9.9; (python_version >= "2.6" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pymongo==3.11.1
python-dateutil==2.8.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
python-magic==0.4.18; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pytz==2020.4; python_version >= "3.7"
requests==2.25.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
six==1.15.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
soupsieve==2.0.1; python_version >= "3.7"
tqdm==4.54.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
tsingspider==1.4.1; python_version >= "3.7"
urllib3==1.26.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.7"
vine==5.0.0; python_version >= "3.6"
wcwidth==0.2.5; python_full_version >= "3.6.1" and python_version >= "3.6"
zipp==3.4.0; python_version < "3.8" and python_version >= "3.6"
